# backend/api/cobytes-openapi.yml
openapi: 3.0.3
info:
  title: Cobytes Security Platform API
  description: |
    The Cobytes Security Platform API provides comprehensive security scanning and reporting capabilities.
    
    ## Authentication
    All endpoints (except health and API info) require JWT authentication.
    Use the `/api/auth/login` endpoint to obtain a token.
    
    ## Base URL
    - Development: `http://localhost:3001`
    - Production: `https://api.cobytes.com`
  version: 1.0.0
  contact:
    name: Cobytes API Support
    email: api-support@cobytes.com
    url: https://docs.cobytes.com
  license:
    name: Proprietary
    url: https://cobytes.com/license
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.cobytes.com
    description: Production server
...
paths:
...
  /api/scans:
    get:
      tags:
        - Scans
      summary: Get All Scans
      description: Get a list of all scans for the authenticated user
      operationId: getAllScans
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          description: Filter by scan status
          schema:
            type: string
            enum: [pending, running, completed, failed]
        - name: limit
          in: query
          description: Number of results per page
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: Pagination offset
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: List of scans
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScansListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Scans
      summary: Create New Scan
      description: Create and start a new security scan
      operationId: createScan
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScanRequest'
      responses:
        '201':
          description: Scan created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/scans/types/available:
    get:
      tags:
        - Scans
      summary: Get Available Scan Types
      description: Get a list of all available scan types
      operationId: getScanTypes
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of scan types
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanTypesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/scans/{scanId}:
    get:
      tags:
        - Scans
      summary: Get Scan Details
      description: Get detailed information about a specific scan
      operationId: getScanById
      security:
        - bearerAuth: []
      parameters:
        - name: scanId
          in: path
          required: true
          description: Scan ID
          schema:
            type: string
      responses:
        '200':
          description: Scan details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanDetailsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Scans
      summary: Delete Scan
      description: Delete a scan and all associated data
      operationId: deleteScan
      security:
        - bearerAuth: []
      parameters:
        - name: scanId
          in: path
          required: true
          description: Scan ID
          schema:
            type: string
      responses:
        '200':
          description: Scan deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateScanRequest:
      type: object
      required:
        - scanType
        - targets
      properties:
        scanType:
          type: string
          enum: [web_vulnerability, network_scan, ssl_tls, dns_security, malware_analysis]
          example: web_vulnerability
        targets:
          type: array
          items:
            type: string
          example: [https://example.com, https://test.example.com]
        configuration:
          type: object
          properties:
            depth:
              type: integer
              example: 3
            threads:
              type: integer
              example: 10
            followRedirects:
              type: boolean
              example: true
            userAgent:
              type: string
              example: Cobytes Security Scanner
        scheduledAt:
          type: string
          format: date-time
    Scan:
      type: object
      properties:
        scanId:
          type: string
          example: scan_1748799588916_dw04msszk
        status:
          type: string
          enum: [pending, running, completed, failed]
          example: pending
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ScanResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Scan'
    ScansListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Scan'
        count:
          type: integer
          example: 1
    ScanType:
      type: object
      properties:
        id:
          type: string
          example: web_vulnerability
        name:
          type: string
          example: Web Vulnerability Scan
        description:
          type: string
          example: Comprehensive web application security scan
        tools:
          type: array
          items:
            type: string
          example: [OWASP ZAP, Burp Suite, Nikto]
    ScanTypesResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/ScanType'
    Vulnerability:
      type: object
      properties:
        severity:
          type: string
          enum: [critical, high, medium, low, info]
          example: high
        title:
          type: string
          example: SQL Injection vulnerability
        description:
          type: string
          example: Potential SQL injection in login form
        cve:
          type: string
          example: CVE-2024-1234
        cwe:
          type: string
          example: CWE-89
        affectedAssets:
          type: array
          items:
            type: string
          example: [192.168.1.100, https://example.com/login]
        recommendation:
          type: string
          example: Use parameterized queries
    ScanResults:
      type: object
      properties:
        vulnerabilities:
          type: array
          items:
            $ref: '#/components/schemas/Vulnerability'
        summary:
          type: object
          properties:
            total:
              type: integer
              example: 2
            high:
              type: integer
              example: 1
            medium:
              type: integer
              example: 1
            low:
              type: integer
              example: 0
    ScanDetailsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          allOf:
            - $ref: '#/components/schemas/Scan'
            - type: object
              properties:
                results:
                  $ref: '#/components/schemas/ScanResults'
...
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Resource deleted successfully
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Error message
        code:
          type: string
          example: ERROR_CODE
        details:
          type: object
          additionalProperties: true
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Validation failed
            code: VALIDATION_ERROR
            details:
              field: Error details
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Invalid or missing authentication
            code: UNAUTHORIZED
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Resource not found
            code: NOT_FOUND
...